refactor: replace fs-extra

+++ src/renames/rename-js.ts
@@ -1,8 +1,9 @@
 import path from 'node:path';
-import fse from 'fs-extra';
+import fse from '@zokugun/fs-extra-plus/async';
+import { type AsyncDResult, err, OK, stringifyError } from '@zokugun/xtry';
 import { globby } from 'globby';
 
-export async function renameJS(dirPath: string, newExtname: string, replace: (string) => string): Promise<void> {
+export async function renameJS(dirPath: string, newExtname: string, replace: (value: string) => string): AsyncDResult {
 	const files = await globby('**/*.js', { cwd: dirPath });
 
 	for(const file of files) {
@@ -10,10 +11,19 @@ export async function renameJS(dirPath: string, newExtname: string, replace: (st
 		const outputFile = inputFile.replace(/\.js$/, newExtname);
 
 		const content = await fse.readFile(inputFile, 'utf8');
-		await fse.writeFile(outputFile, replace(content), 'utf8');
+		if(content.fails) {
+			return err(stringifyError(content.error));
+		}
+
+		await fse.writeFile(outputFile, replace(content.value), 'utf8');
 
 		if(outputFile !== inputFile) {
-			await fse.unlink(inputFile);
+			const result = await fse.unlink(inputFile);
+			if(result.fails) {
+				return err(stringifyError(result.error));
+			}
 		}
 	}
+
+	return OK;
 }
